requirements.txt
flask
matplotlib
tensorflow
google-cloud-storage
gunicorn
Pillow

#jalankan dengan perintah "pip install -r requirements.txt"